Aggregator for the shards
