+++ backend/shard/src/node.rs
@@ -1,242 +1,213 @@
+use std::collections::BTreeMap;
+use std::net::Ipv4Addr;
+use std::time::{Duration, Instant};
+
+use crate::aggregator::{AddNode, Aggregator};
+// use crate::chain::{Chain, RemoveNode, UpdateNode};
+use actix::prelude::*;
+use actix_web_actors::ws::{self, CloseReason};
 use bytes::Bytes;
-use std::sync::Arc;
+use shared::node::{NodeMessage, Payload};
+use shared::types::{ConnId, NodeId};
+use shared::ws::{MultipartHandler, WsMessage};
 
-use shared::types::{
-    Block, BlockDetails, NodeDetails, NodeHardware, NodeIO, NodeId, NodeLocation, NodeStats,
-    Timestamp,
-};
-use shared::util::now;
-use shared::node::SystemInterval;
+/// How often heartbeat pings are sent
+const HEARTBEAT_INTERVAL: Duration = Duration::from_secs(20);
+/// How long before lack of client response causes a timeout
+const CLIENT_TIMEOUT: Duration = Duration::from_secs(60);
 
-pub mod connector;
-
-/// Minimum time between block below broadcasting updates to the browser gets throttled, in ms.
-const THROTTLE_THRESHOLD: u64 = 100;
-/// Minimum time of intervals for block updates sent to the browser when throttled, in ms.
-const THROTTLE_INTERVAL: u64 = 1000;
-
-pub struct Node {
-    /// Static details
-    details: NodeDetails,
-    /// Basic stats
-    stats: NodeStats,
-    /// Node IO stats
-    io: NodeIO,
-    /// Best block
-    best: BlockDetails,
-    /// Finalized block
-    finalized: Block,
-    /// Timer for throttling block updates
-    throttle: u64,
-    /// Hardware stats over time
-    hardware: NodeHardware,
-    /// Physical location details
-    location: Option<Arc<NodeLocation>>,
-    /// Flag marking if the node is stale (not syncing or producing blocks)
-    stale: bool,
-    /// Unix timestamp for when node started up (falls back to connection time)
-    startup_time: Option<Timestamp>,
-    /// Network state
-    network_state: Option<Bytes>,
+pub struct NodeConnector {
+    /// Multiplexing connections by id
+    multiplex: BTreeMap<ConnId, ConnMultiplex>,
+    /// Client must send ping at least once every 60 seconds (CLIENT_TIMEOUT),
+    hb: Instant,
+    /// Aggregator actor address
+    aggregator: Addr<Aggregator>,
+    /// IP address of the node this connector is responsible for
+    ip: Option<Ipv4Addr>,
+    /// Helper for handling continuation messages
+    multipart: MultipartHandler,
 }
 
-impl Node {
-    pub fn new(mut details: NodeDetails) -> Self {
-        let startup_time = details
-            .startup_time
-            .take()
-            .and_then(|time| time.parse().ok());
+enum ConnMultiplex {
+    Connected {
+        /// Id of the node this multiplex connector is responsible for handling
+        nid: NodeId,
+        // /// Chain address to which this multiplex connector is delegating messages
+        // chain: Addr<Chain>,
+    },
+    Waiting {
+        /// Backlog of messages to be sent once we get a recipient handle to the chain
+        backlog: Vec<Payload>,
+    },
+}
 
-        Node {
-            details,
-            stats: NodeStats::default(),
-            io: NodeIO::default(),
-            best: BlockDetails::default(),
-            finalized: Block::zero(),
-            throttle: 0,
-            hardware: NodeHardware::default(),
-            location: None,
-            stale: false,
-            startup_time,
-            network_state: None,
+impl Default for ConnMultiplex {
+    fn default() -> Self {
+        ConnMultiplex::Waiting {
+            backlog: Vec::new(),
+        }
+    }
+}
+
+impl Actor for NodeConnector {
+    type Context = ws::WebsocketContext<Self>;
+
+    fn started(&mut self, ctx: &mut Self::Context) {
+        self.heartbeat(ctx);
+    }
+
+    fn stopped(&mut self, _: &mut Self::Context) {
+        // for mx in self.multiplex.values() {
+        //     if let ConnMultiplex::Connected { chain, nid } = mx {
+        //         chain.do_send(RemoveNode(*nid));
+        //     }
+        // }
+    }
+}
+
+impl NodeConnector {
+    pub fn new(aggregator: Addr<Aggregator>, ip: Option<Ipv4Addr>) -> Self {
+        Self {
+            multiplex: BTreeMap::new(),
+            hb: Instant::now(),
+            aggregator,
+            ip,
+            multipart: MultipartHandler::default(),
         }
     }
 
-    pub fn details(&self) -> &NodeDetails {
-        &self.details
+    fn heartbeat(&self, ctx: &mut <Self as Actor>::Context) {
+        ctx.run_interval(HEARTBEAT_INTERVAL, |act, ctx| {
+            // check client heartbeats
+            if Instant::now().duration_since(act.hb) > CLIENT_TIMEOUT {
+                // stop actor
+                ctx.close(Some(CloseReason {
+                    code: ws::CloseCode::Abnormal,
+                    description: Some("Missed heartbeat".into()),
+                }));
+                ctx.stop();
+            }
+        });
     }
 
-    pub fn stats(&self) -> &NodeStats {
-        &self.stats
-    }
-
-    pub fn io(&self) -> &NodeIO {
-        &self.io
-    }
-
-    pub fn best(&self) -> &Block {
-        &self.best.block
-    }
-
-    pub fn best_timestamp(&self) -> u64 {
-        self.best.block_timestamp
-    }
-
-    pub fn finalized(&self) -> &Block {
-        &self.finalized
-    }
-
-    pub fn hardware(&self) -> &NodeHardware {
-        &self.hardware
-    }
-
-    pub fn location(&self) -> Option<&NodeLocation> {
-        self.location.as_deref()
-    }
-
-    pub fn update_location(&mut self, location: Arc<NodeLocation>) {
-        self.location = Some(location);
-    }
-
-    pub fn block_details(&self) -> &BlockDetails {
-        &self.best
-    }
-
-    pub fn update_block(&mut self, block: Block) -> bool {
-        if block.height > self.best.block.height {
-            self.stale = false;
-            self.best.block = block;
-
-            true
-        } else {
-            false
-        }
-    }
-
-    pub fn update_details(
+    fn handle_message(
         &mut self,
-        timestamp: u64,
-        propagation_time: Option<u64>,
-    ) -> Option<&BlockDetails> {
-        self.best.block_time = timestamp - self.best.block_timestamp;
-        self.best.block_timestamp = timestamp;
-        self.best.propagation_time = propagation_time;
+        msg: NodeMessage,
+        data: Bytes,
+        ctx: &mut <Self as Actor>::Context,
+    ) {
+        let conn_id = msg.id();
+        let payload = msg.into();
 
-        if self.throttle < timestamp {
-            if self.best.block_time <= THROTTLE_THRESHOLD {
-                self.throttle = timestamp + THROTTLE_INTERVAL;
+        match self.multiplex.entry(conn_id).or_default() {
+            ConnMultiplex::Connected { nid } => {
+                // chain.do_send(UpdateNode {
+                //     nid: *nid,
+                //     raw: Some(data),
+                //     payload,
+                // });
             }
+            ConnMultiplex::Waiting { backlog } => {
+                if let Payload::SystemConnected(connected) = payload {
+                    println!("Node connected {:?}", connected);
+                    self.aggregator.do_send(AddNode {
+                        genesis_hash: connected.genesis_hash,
+                        ip: self.ip,
+                        // node: connected.node,
+                        // conn_id,
+                        // node_connector: ctx.address(),
+                    });
+                } else {
+                    if backlog.len() >= 10 {
+                        backlog.remove(0);
+                    }
 
-            Some(&self.best)
-        } else {
-            None
-        }
-    }
-
-    pub fn update_hardware(&mut self, interval: &SystemInterval) -> bool {
-        let mut changed = false;
-
-        if let Some(upload) = interval.bandwidth_upload {
-            changed |= self.hardware.upload.push(upload);
-        }
-        if let Some(download) = interval.bandwidth_download {
-            changed |= self.hardware.download.push(download);
-        }
-        self.hardware.chart_stamps.push(now() as f64);
-
-        changed
-    }
-
-    pub fn update_stats(&mut self, interval: &SystemInterval) -> Option<&NodeStats> {
-        let mut changed = false;
-
-        if let Some(peers) = interval.peers {
-            if peers != self.stats.peers {
-                self.stats.peers = peers;
-                changed = true;
+                    backlog.push(payload);
+                }
             }
         }
-        if let Some(txcount) = interval.txcount {
-            if txcount != self.stats.txcount {
-                self.stats.txcount = txcount;
-                changed = true;
+    }
+}
+
+// #[derive(Message)]
+// #[rtype(result = "()")]
+// pub struct Initialize {
+//     pub nid: NodeId,
+//     pub conn_id: ConnId,
+//     pub chain: Addr<Chain>,
+// }
+
+// impl Handler<Initialize> for NodeConnector {
+//     type Result = ();
+
+//     fn handle(&mut self, msg: Initialize, _: &mut Self::Context) {
+//         let Initialize {
+//             nid,
+//             conn_id,
+//             chain,
+//         } = msg;
+//         log::trace!(target: "NodeConnector::Initialize", "Initializing a node, nid={}, on conn_id={}", nid, conn_id);
+//         let mx = self.multiplex.entry(conn_id).or_default();
+
+//         if let ConnMultiplex::Waiting { backlog } = mx {
+//             for payload in backlog.drain(..) {
+//                 chain.do_send(UpdateNode {
+//                     nid,
+//                     raw: None,
+//                     payload,
+//                 });
+//             }
+
+//             *mx = ConnMultiplex::Connected {
+//                 nid,
+//                 chain: chain.clone(),
+//             };
+//         };
+
+//         // Acquire the node's physical location
+//         if let Some(ip) = self.ip {
+//             let _ = self.locator.do_send(LocateRequest { ip, nid, chain });
+//         }
+//     }
+// }
+
+impl StreamHandler<Result<ws::Message, ws::ProtocolError>> for NodeConnector {
+    fn handle(&mut self, msg: Result<ws::Message, ws::ProtocolError>, ctx: &mut Self::Context) {
+        self.hb = Instant::now();
+
+        let data = match msg.map(|msg| self.multipart.handle(msg)) {
+            Ok(WsMessage::Nop) => return,
+            Ok(WsMessage::Ping(msg)) => {
+                ctx.pong(&msg);
+                return;
             }
+            Ok(WsMessage::Data(data)) => data,
+            Ok(WsMessage::Close(reason)) => {
+                ctx.close(reason);
+                ctx.stop();
+                return;
+            }
+            Err(error) => {
+                log::error!("{:?}", error);
+                ctx.stop();
+                return;
+            }
+        };
+
+        match serde_json::from_slice(&data) {
+            Ok(msg) => self.handle_message(msg, data, ctx),
+            #[cfg(debug)]
+            Err(err) => {
+                let data: &[u8] = data.get(..512).unwrap_or_else(|| &data);
+                log::warn!(
+                    "Failed to parse node message: {} {}",
+                    err,
+                    std::str::from_utf8(data).unwrap_or_else(|_| "INVALID UTF8")
+                )
+            }
+            #[cfg(not(debug))]
+            Err(_) => (),
         }
-
-        if changed {
-            Some(&self.stats)
-        } else {
-            None
-        }
-    }
-
-    pub fn update_io(&mut self, interval: &SystemInterval) -> Option<&NodeIO> {
-        let mut changed = false;
-
-        if let Some(size) = interval.used_state_cache_size {
-            changed |= self.io.used_state_cache_size.push(size);
-        }
-
-        if changed {
-            Some(&self.io)
-        } else {
-            None
-        }
-    }
-
-    pub fn update_finalized(&mut self, block: Block) -> Option<&Block> {
-        if block.height > self.finalized.height {
-            self.finalized = block;
-            Some(self.finalized())
-        } else {
-            None
-        }
-    }
-
-    pub fn update_stale(&mut self, threshold: u64) -> bool {
-        if self.best.block_timestamp < threshold {
-            self.stale = true;
-        }
-
-        self.stale
-    }
-
-    pub fn stale(&self) -> bool {
-        self.stale
-    }
-
-    pub fn set_validator_address(&mut self, addr: Box<str>) {
-        self.details.validator = Some(addr);
-    }
-
-    pub fn set_network_state(&mut self, state: Bytes) {
-        self.network_state = Some(state);
-    }
-
-    pub fn network_state(&self) -> Option<Bytes> {
-        use serde::Deserialize;
-        use serde_json::value::RawValue;
-
-        #[derive(Deserialize)]
-        struct Wrapper<'a> {
-            #[serde(borrow)]
-            #[serde(alias = "network_state")]
-            state: &'a RawValue,
-        }
-
-        let raw = self.network_state.as_ref()?;
-        let wrap: Wrapper = serde_json::from_slice(raw).ok()?;
-        let json = wrap.state.get();
-
-        // Handle old nodes that exposed network_state as stringified JSON
-        if let Ok(stringified) = serde_json::from_str::<String>(json) {
-            Some(stringified.into())
-        } else {
-            Some(json.to_owned().into())
-        }
-    }
-
-    pub fn startup_time(&self) -> Option<Timestamp> {
-        self.startup_time
     }
 }
